moved code to new app 'mapbender_compatibility'
